Initial Contact US form refactoring commit
